RUM-6286: post-review fixes - moving StringBuilderExt into internal module
