+++ dd-sdk-android-internal/src/test/java/com/datadog/internal/utils/StringBuilderExtKtTest.kt
@@ -1,5 +1,12 @@
-package com.datadog.android.core
+/*
+ * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+ * This product includes software developed at Datadog (https://www.datadoghq.com/).
+ * Copyright 2016-Present Datadog, Inc.
+ */
 
+package com.datadog.internal.utils
+
+import com.datadog.android.internal.utils.appendIfNotEmpty
 import fr.xgouchet.elmyr.annotation.StringForgery
 import fr.xgouchet.elmyr.junit5.ForgeExtension
 import org.assertj.core.api.Assertions.assertThat
@@ -38,7 +45,6 @@ internal class StringBuilderExtKtTest {
         buffer.appendIfNotEmpty(" ")
 
         // Then
-
         assertThat(buffer.toString()).isEqualTo("$initialContent ")
     }
 
@@ -51,7 +57,6 @@ internal class StringBuilderExtKtTest {
         buffer.appendIfNotEmpty(' ')
 
         // Then
-
         assertThat(buffer.toString()).isEqualTo("")
     }
 
@@ -64,7 +69,6 @@ internal class StringBuilderExtKtTest {
         buffer.appendIfNotEmpty(" ")
 
         // Then
-
         assertThat(buffer.toString()).isEqualTo("")
     }
 }
