+++ features/dd-sdk-android-rum/src/main/kotlin/com/datadog/android/rum/internal/anr/AndroidTraceParser.kt
@@ -7,8 +7,8 @@
 package com.datadog.android.rum.internal.anr
 
 import com.datadog.android.api.InternalLogger
-import com.datadog.android.core.appendIfNotEmpty
 import com.datadog.android.core.feature.event.ThreadDump
+import com.datadog.android.internal.utils.appendIfNotEmpty
 import java.io.IOException
 import java.io.InputStream
 import java.util.Locale
